Rewrite gemmini conv with user-level schedule ops (#613)
This PR rewrites the previous Gemmini conv artifact evaluation code to
use user-level scheduling ops we defined for Gemmini matmul, such as
`fission_as_much_as_possible` and `fuse_all_loops`. There are many TODOs
to clean up the Gemmini code, such as 1) write a parameterized schedule
that works for various resnet matmul and conv sizes (can we clean up
`sched_one_block` and somehow clean up the loop reorganization code?),
2) run FireSim to check the performance. New schedules might achieve
better performance than the previous one.
